requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==4.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
wrapt==1.17.2
